Server fails to start because of two services with the same name. Merge them into one file.
