+++ core/src/main/java/inetsoft/web/AutoSaveService.java
@@ -12,18 +12,36 @@
  * person.
  */
 package inetsoft.web;
-import org.springframework.stereotype.Service;
+
+import inetsoft.analytic.composition.ViewsheetService;
+import inetsoft.cluster.*;
+import inetsoft.report.composition.WorksheetEngine;
+import inetsoft.report.composition.event.AssetEventUtil;
+import inetsoft.sree.internal.SUtil;
+import inetsoft.sree.security.IdentityID;
 import inetsoft.storage.BlobStorage;
-import org.springframework.scheduling.annotation.Scheduled;
+import inetsoft.uql.asset.*;
+import inetsoft.uql.asset.internal.AssetUtil;
+import inetsoft.util.audit.ActionRecord;
+import inetsoft.util.audit.Audit;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
+
 import java.io.FileNotFoundException;
+import java.security.Principal;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
 
 @Service
 @Lazy(false)
+@ClusterProxy
 public class AutoSaveService {
+   public AutoSaveService(ViewsheetService viewsheetService) {
+      this.viewsheetService = viewsheetService;
+   }
+
    @Scheduled(fixedRate = 10800000L)
    public void removeExpiredAutoSaveFiles() {
       BlobStorage<AutoSaveUtils.Metadata> blobStorage = AutoSaveUtils.getStorage(null);
@@ -36,7 +54,7 @@ public class AutoSaveService {
       List<String> autoSavedFiles = AutoSaveUtils.getAutoSavedFiles(null, isRecycle);
       long assetLastModifiedTime;
 
-      for(String fileName: autoSavedFiles) {
+      for(String fileName : autoSavedFiles) {
          String file = AutoSaveUtils.getAutoSavedByName(fileName, isRecycle);
 
          try {
@@ -53,4 +71,34 @@ public class AutoSaveService {
          }
       }
    }
+
+   @ClusterProxyMethod(WorksheetEngine.CACHE_NAME)
+   public Boolean restoreAutoSaveAssets(@ClusterProxyKey String id, String assetName, boolean override,
+                                        Principal principal)
+      throws Exception
+   {
+      // Get auto save sheet from engine.
+      AssetEntry entry = AutoSaveUtils.createAssetEntry(id);
+      AssetRepository repository = AssetUtil.getAssetRepository(false);
+      AbstractSheet sheet = repository.getSheet(entry, principal, false, AssetContent.ALL);
+      IdentityID pId = IdentityID.getIdentityIDFromKey(principal.getName());
+
+      // Save auto save sheet to engine.
+      AssetEntry.Type type = id.startsWith("8^VIEWSHEET") ? AssetEntry.Type.VIEWSHEET :
+         AssetEntry.Type.WORKSHEET;
+      AssetEntry nentry = new AssetEntry(AssetRepository.GLOBAL_SCOPE, type, assetName,
+                                         pId);
+
+      if(!override && viewsheetService.isDuplicatedEntry(repository, nentry)) {
+         return false;
+      }
+
+      repository.setSheet(nentry, sheet, principal, false);
+      ActionRecord actionRecord = SUtil.getActionRecord(principal, ActionRecord.ACTION_NAME_CREATE,
+                                                        assetName, AssetEventUtil.getObjectType(entry));
+      Audit.getInstance().auditAction(actionRecord, principal);
+      return true;
+   }
+
+   private final ViewsheetService viewsheetService;
 }
